// frontend/src/Models/Reviews.ts
export interface Review {
  id: string;
  promptId: string;
  userId: string;
  userName: string;
  rating: number;
  comment: string;     
}

export interface ReviewsApiResponse {
  content: Review[];
  pageable: {
    pageNumber: number;
    pageSize: number;
    sort: {
      empty: boolean;
      sorted: boolean;
      unsorted: boolean;
    };
    offset: number;
    paged: boolean;
    unpaged: boolean;
  };
  last: boolean;
  totalElements: number;
  totalPages: number;
  size: number;
  number: number;
  sort: {
    empty: boolean;
    sorted: boolean;
    unsorted: boolean;
  };
  first: boolean;
  numberOfElements: number;
  empty: boolean;
}